myproject-django-todo/requirements.txt
ansicon==1.89.0
asgiref==3.9.1
awsebcli==3.23.3
bcrypt==4.3.0
blessed==1.21.0
blinker==1.9.0
boto3==1.40.4
botocore==1.40.4
cement==2.10.14
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
Deprecated==1.2.18
distlib==0.4.0
Django==4.2.23
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
fabric==3.2.2
filelock==3.18.0
Flask==3.1.1
idna==3.10
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jinxed==1.3.0
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
pypiwin32==223
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2025.2
pywin32==311
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.1
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
tzdata==2025.2
urllib3==1.26.20
virtualenv==20.32.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
